feat: Alterações gerais na arquitetura e mensagens de erro em constantes

+++ src/main/java/com/uniacademia/enade/api/controllers/CategoryController.java
@@ -1,6 +1,8 @@
 package com.uniacademia.enade.api.controllers;
 
 import java.security.NoSuchAlgorithmException;
+import java.util.List;
+import java.util.Optional;
 
 import javax.validation.Valid;
 
@@ -13,6 +15,7 @@ import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -21,8 +24,10 @@ import org.springframework.web.bind.annotation.RestController;
 import com.uniacademia.enade.api.dto.EditCategory;
 import com.uniacademia.enade.api.dto.IncludeCategory;
 import com.uniacademia.enade.api.entity.Category;
+import com.uniacademia.enade.api.enumerator.GenericMessages;
 import com.uniacademia.enade.api.response.Response;
 import com.uniacademia.enade.api.service.CategoryService;
+import com.uniacademia.enade.api.utils.Messages;
 
 @RestController
 @CrossOrigin(origins = "*")
@@ -38,10 +43,11 @@ public class CategoryController {
 	}
 
 	@GetMapping("/all")
-	public ResponseEntity<Response<String>> all() throws NoSuchAlgorithmException {
-		Response<String> response = new Response<String>();
+	public ResponseEntity<Response<List<Category>>> all() throws NoSuchAlgorithmException {
+		Response<List<Category>> response = new Response<List<Category>>();
 
-		response.setData("Batata");
+		List<Category> categories = categoryService.findAll();
+		response.setData(categories);
 
 		return ResponseEntity.ok(response);
 	}
@@ -49,11 +55,11 @@ public class CategoryController {
 	@PostMapping("/include")
 	public ResponseEntity<Response<Category>> include(@Valid @RequestBody IncludeCategory includeCategory,
 			BindingResult result) throws NoSuchAlgorithmException {
-		log.info("Cadastrando Categoria: {}", includeCategory.toString());
+		log.info("Iniciando cadastro da categoria: {}", includeCategory.toString());
 		Response<Category> response = new Response<Category>();
 
 		if (result.hasErrors()) {
-			log.error("Erro validando dados de cadastro de Categoria: {}", result.getAllErrors());
+			log.error("Erro validando dados de cadastro de categoria: {}", result.getAllErrors());
 			result.getAllErrors().forEach(error -> response.addFieldError(error.getDefaultMessage()));
 
 			return ResponseEntity.badRequest().body(response);
@@ -69,11 +75,11 @@ public class CategoryController {
 	@PatchMapping("/edit")
 	public ResponseEntity<Response<Category>> edit(@Valid @RequestBody EditCategory editCategory, BindingResult result)
 			throws NoSuchAlgorithmException {
-		log.info("Editar Categoria: {}", editCategory.toString());
+		log.info("Iniciando edição da categoria: {}", editCategory.toString());
 		Response<Category> response = new Response<Category>();
 
 		if (result.hasErrors()) {
-			log.error("Erro validando dados para edição de Categoria: {}", result.getAllErrors());
+			log.error("Erro validando dados para edição de categoria: {}", result.getAllErrors());
 			result.getAllErrors().forEach(error -> response.addFieldError(error.getDefaultMessage()));
 
 			return ResponseEntity.badRequest().body(response);
@@ -86,14 +92,22 @@ public class CategoryController {
 		return ResponseEntity.ok(response);
 	}
 
-	@DeleteMapping("/remove")
-	public ResponseEntity<Response<Long>> remove(@Valid @RequestBody Long id, BindingResult result)
-			throws NoSuchAlgorithmException {
-		log.info("Removendo Categoria: {}", id);
-		Response<Long> response = new Response<Long>();
-		this.categoryService.deleteById(id);
+	@DeleteMapping("/remove/{id}")
+	public ResponseEntity<Response<Category>> remove(@PathVariable("id") Long id) throws NoSuchAlgorithmException {
+		log.info("Iniciando remoção da categoria: {}", id);
+		Response<Category> response = new Response<Category>();
 
-		response.setData(id);
+		Optional<Category> category = this.categoryService.findById(id);
+		if (category.isEmpty()) {
+			log.info("Erro ao validar o 'Id' para remoção da categoria: {}", id);
+			response.addError(Messages.getCategoriaError(GenericMessages.NONEXISTENT.toString()));
+
+			return ResponseEntity.badRequest().body(response);
+		}
+
+		this.categoryService.deleteById(id);
+		response.setData(category.get());
+
 		return ResponseEntity.ok(response);
 	}
 }
